refactor: use the new domain service to close subscription in portal api
