+++ gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/pom.xml
@@ -241,6 +241,13 @@
 			<artifactId>wiremock</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>io.gravitee.apim.rest.api</groupId>
+			<artifactId>gravitee-apim-rest-api-service</artifactId>
+			<version>${project.version}</version>
+			<type>test-jar</type>
+			<scope>test</scope>
+		</dependency>
 
 		<!-- Json Path Find-->
 		<dependency>
