+++ gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/test/java/io/gravitee/rest/api/portal/rest/resource/AbstractResourceTest.java
@@ -15,11 +15,9 @@
  */
 package io.gravitee.rest.api.portal.rest.resource;
 
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.reset;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.gravitee.definition.jackson.datatype.GraviteeMapper;
 import io.gravitee.rest.api.portal.rest.JerseySpringTest;
 import io.gravitee.rest.api.portal.rest.mapper.AnalyticsMapper;
 import io.gravitee.rest.api.portal.rest.mapper.ApiMapper;
@@ -39,6 +37,7 @@ import io.gravitee.rest.api.portal.rest.mapper.SubscriptionMapper;
 import io.gravitee.rest.api.portal.rest.mapper.ThemeMapper;
 import io.gravitee.rest.api.portal.rest.mapper.TicketMapper;
 import io.gravitee.rest.api.portal.rest.mapper.UserMapper;
+import io.gravitee.rest.api.portal.rest.spring.ResourceContextConfiguration;
 import io.gravitee.rest.api.security.authentication.AuthenticationProvider;
 import io.gravitee.rest.api.security.authentication.AuthenticationProviderManager;
 import io.gravitee.rest.api.security.cookies.CookieGenerator;
@@ -61,7 +60,6 @@ import io.gravitee.rest.api.service.GroupService;
 import io.gravitee.rest.api.service.HealthCheckService;
 import io.gravitee.rest.api.service.IdentityService;
 import io.gravitee.rest.api.service.LogsService;
-import io.gravitee.rest.api.service.MediaService;
 import io.gravitee.rest.api.service.MembershipService;
 import io.gravitee.rest.api.service.MessageService;
 import io.gravitee.rest.api.service.NotifierService;
@@ -104,12 +102,8 @@ import java.util.Optional;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.support.AnnotationConfigContextLoader;
 
 /**
  * @author David BRASSELY (david.brassely at graviteesource.com)
@@ -117,7 +111,7 @@ import org.springframework.test.context.support.AnnotationConfigContextLoader;
  * @author Florent CHAMFROY (florent.chamfroy at graviteesource.com)
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(loader = AnnotationConfigContextLoader.class)
+@ContextConfiguration(classes = { ResourceContextConfiguration.class })
 public abstract class AbstractResourceTest extends JerseySpringTest {
 
     @Autowired
@@ -422,366 +416,6 @@ public abstract class AbstractResourceTest extends JerseySpringTest {
         reset(accessControlService);
         reset(themeService);
         reset(themeMapper);
-    }
-
-    @Configuration
-    @PropertySource("classpath:/io/gravitee/rest/api/portal/rest/resource/jwt.properties")
-    static class ContextConfiguration {
-
-        @Bean
-        public ApiService apiService() {
-            return mock(ApiService.class);
-        }
-
-        @Bean
-        public ApiSearchService apiSearchService() {
-            return mock(ApiSearchService.class);
-        }
-
-        @Bean
-        public ApiAuthorizationService apiAuthorizationService() {
-            return mock(ApiAuthorizationService.class);
-        }
-
-        @Bean
-        public ApiEntrypointService apiEntrypointService() {
-            return mock(ApiEntrypointService.class);
-        }
-
-        @Bean
-        public ApplicationService applicationService() {
-            return mock(ApplicationService.class);
-        }
-
-        @Bean
-        public ApplicationTypeService applicationTypeService() {
-            return mock(ApplicationTypeService.class);
-        }
-
-        @Bean
-        public UserService userService() {
-            return mock(UserService.class);
-        }
-
-        @Bean
-        public PolicyService policyService() {
-            return mock(PolicyService.class);
-        }
-
-        @Bean
-        public FetcherService fetcherService() {
-            return mock(FetcherService.class);
-        }
-
-        @Bean
-        public SwaggerService swaggerService() {
-            return mock(SwaggerService.class);
-        }
-
-        @Bean
-        public MembershipService membershipService() {
-            return mock(MembershipService.class);
-        }
-
-        @Bean
-        public RoleService roleService() {
-            return mock(RoleService.class);
-        }
-
-        @Bean("oauth2")
-        public AuthenticationProvider authenticationProvider() {
-            return mock(AuthenticationProvider.class);
-        }
-
-        @Bean
-        public PageService pageService() {
-            return mock(PageService.class);
-        }
-
-        @Bean
-        public GroupService groupService() {
-            return mock(GroupService.class);
-        }
-
-        @Bean
-        public RatingService ratingService() {
-            return mock(RatingService.class);
-        }
-
-        @Bean
-        public PermissionService permissionService() {
-            return mock(PermissionService.class);
-        }
-
-        @Bean
-        public NotifierService notifierService() {
-            return mock(NotifierService.class);
-        }
-
-        @Bean
-        public TopApiService topApiService() {
-            return mock(TopApiService.class);
-        }
-
-        @Bean
-        public CookieGenerator jwtCookieGenerator() {
-            return mock(CookieGenerator.class);
-        }
-
-        @Bean
-        public TaskService taskService() {
-            return mock(TaskService.class);
-        }
-
-        @Bean
-        public LogsService logsService() {
-            return mock(LogsService.class);
-        }
-
-        @Bean
-        public QualityMetricsService qualityMetricsService() {
-            return mock(QualityMetricsService.class);
-        }
-
-        @Bean
-        public MessageService messageService() {
-            return mock(MessageService.class);
-        }
-
-        @Bean
-        public SocialIdentityProviderService socialIdentityProviderService() {
-            return mock(SocialIdentityProviderService.class);
-        }
-
-        @Bean
-        public TagService tagService() {
-            return mock(TagService.class);
-        }
-
-        @Bean
-        public MediaService mediaService() {
-            return mock(MediaService.class);
-        }
-
-        @Bean
-        public ParameterService parameterService() {
-            return mock(ParameterService.class);
-        }
-
-        @Bean
-        public ApiMetadataService metadataService() {
-            return mock(ApiMetadataService.class);
-        }
-
-        @Bean
-        public PlanService planService() {
-            return mock(PlanService.class);
-        }
-
-        @Bean
-        public PlanSearchService planSearchService() {
-            return mock(PlanSearchService.class);
-        }
-
-        @Bean
-        public SubscriptionService subscriptionService() {
-            return mock(SubscriptionService.class);
-        }
-
-        @Bean
-        public EntrypointService entrypointService() {
-            return mock(EntrypointService.class);
-        }
-
-        @Bean
-        public ApiKeyService apiKeyService() {
-            return mock(ApiKeyService.class);
-        }
-
-        @Bean
-        public AnalyticsService analyticsService() {
-            return mock(AnalyticsService.class);
-        }
-
-        @Bean
-        public PortalNotificationConfigService portalNotificationConfigService() {
-            return mock(PortalNotificationConfigService.class);
-        }
-
-        @Bean
-        public PortalNotificationService portalNotificationService() {
-            return mock(PortalNotificationService.class);
-        }
-
-        @Bean
-        public GenericNotificationConfigService genericNotificationConfigService() {
-            return mock(GenericNotificationConfigService.class);
-        }
-
-        @Bean
-        public CategoryService categoryService() {
-            return mock(CategoryService.class);
-        }
-
-        @Bean
-        public ApiCategoryService apiCategoryService() {
-            return mock(ApiCategoryService.class);
-        }
-
-        @Bean
-        public TicketService ticketService() {
-            return mock(TicketService.class);
-        }
-
-        @Bean
-        public ConfigService configService() {
-            return mock(ConfigService.class);
-        }
-
-        @Bean
-        public ApiMapper apiMapper() {
-            return mock(ApiMapper.class);
-        }
-
-        @Bean
-        public PageMapper pageMapper() {
-            return mock(PageMapper.class);
-        }
-
-        @Bean
-        public PlanMapper planMapper() {
-            return mock(PlanMapper.class);
-        }
-
-        @Bean
-        public RatingMapper ratingMapper() {
-            return mock(RatingMapper.class);
-        }
-
-        @Bean
-        public SubscriptionMapper subscriptionMapper() {
-            return mock(SubscriptionMapper.class);
-        }
-
-        @Bean
-        public KeyMapper keyMapper() {
-            return mock(KeyMapper.class);
-        }
-
-        @Bean
-        public ApplicationMapper applicationMapper() {
-            return mock(ApplicationMapper.class);
-        }
-
-        @Bean
-        public MemberMapper memberMapper() {
-            return mock(MemberMapper.class);
-        }
-
-        @Bean
-        public UserMapper userMapper() {
-            return mock(UserMapper.class);
-        }
-
-        @Bean
-        public LogMapper logMapper() {
-            return mock(LogMapper.class);
-        }
-
-        @Bean
-        public AnalyticsMapper analyticsMapper() {
-            return mock(AnalyticsMapper.class);
-        }
-
-        @Bean
-        public PortalNotificationMapper portalNotificationMapper() {
-            return mock(PortalNotificationMapper.class);
-        }
-
-        @Bean
-        public CategoryMapper categoryMapper() {
-            return mock(CategoryMapper.class);
-        }
-
-        @Bean
-        public TicketMapper ticketMapper() {
-            return mock(TicketMapper.class);
-        }
-
-        @Bean
-        public ConfigurationMapper configMapper() {
-            return mock(ConfigurationMapper.class);
-        }
-
-        @Bean
-        public IdentityProviderMapper identityProviderMapper() {
-            return mock(IdentityProviderMapper.class);
-        }
-
-        @Bean
-        public HealthCheckService healthCheckService() {
-            return mock(HealthCheckService.class);
-        }
-
-        @Bean
-        public IdentityService identityService() {
-            return mock(IdentityService.class);
-        }
-
-        @Bean
-        public FilteringService filteringService() {
-            return mock(FilteringService.class);
-        }
-
-        @Bean
-        public ApplicationMetadataService applicationMetadataService() {
-            return mock(ApplicationMetadataService.class);
-        }
-
-        @Bean
-        public ReferenceMetadataMapper referenceMetadataMapper() {
-            return mock(ReferenceMetadataMapper.class);
-        }
-
-        @Bean
-        public CustomUserFieldService customUserFieldService() {
-            return mock(CustomUserFieldService.class);
-        }
-
-        @Bean
-        public IdentityProviderActivationService identityProviderActivationService() {
-            return mock(IdentityProviderActivationService.class);
-        }
-
-        @Bean
-        public AuthoritiesProvider authoritiesProvider() {
-            return mock(AuthoritiesProvider.class);
-        }
-
-        @Bean
-        public EnvironmentService environmentService() {
-            return mock(EnvironmentService.class);
-        }
-
-        @Bean
-        public AccessControlService accessControlService() {
-            return mock(AccessControlService.class);
-        }
-
-        @Bean
-        public ThemeService themeService() {
-            return mock(ThemeService.class);
-        }
-
-        @Bean
-        public ThemeMapper themeMapper() {
-            return mock(ThemeMapper.class);
-        }
-
-        @Bean
-        public ObjectMapper objectMapper() {
-            return new GraviteeMapper();
-        }
     }
 
     @Priority(50)
